+++ crates/egui_flex/CHANGELOG.md
@@ -1,5 +1,12 @@
 # egui_flex changelog
 
+## Unreleased
+
+- Add ways to specify the flex container size, via `Flex::width` and related methods
+- Implement `Flex::justify`
+- Change `Flex::align_content` default to stretch to match the css flexbox behavior
+- Improved textedit flex widget behaviour (once https://github.com/emilk/egui/pull/5275 is merged / released)
+
 ## 0.1.1
 
 - Add comment about not all `FlexAlignContent` variants being implemented